Fix type casting for MaxReadSize assignment
Explicitly cast MemDMA.MAX_READ_SIZE to int when assigning
to ScatterReadMap.MaxReadSize to ensure type correctness
and prevent potential runtime issues.

+++ src/DMA/MemoryInterface.cs
@@ -15,7 +15,7 @@ namespace eft_dma_radar.DMA
         /// </summary>
         public static void ModuleInit()
         {
-            ScatterReadMap.MaxReadSize = MemDMA.MAX_READ_SIZE;
+            ScatterReadMap.MaxReadSize = (int)MemDMA.MAX_READ_SIZE;
             Memory = new MemDMA();
             Debug.WriteLine("DMA Initialized!");
         }
